# webdoc/docs/project_specs/007-api-layer-crud/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Focal CRUD API
  description: Configuration management endpoints for Focal cognitive engine
  version: 1.0.0
servers:
  - url: /v1
    description: API v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT with tenant_id, sub (user_id), and roles claims
  schemas:
...
    PaginatedResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean

    # Agents
    AgentSettings:
      type: object
      properties:
        llm_provider:
          type: string
          example: anthropic
        llm_model:
          type: string
          example: claude-3-5-sonnet
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 0.7
        max_tokens:
          type: integer
          minimum: 1
          default: 1024
    Agent:
      type: object
      required: [id, name, enabled, current_version, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
        enabled:
          type: boolean
        current_version:
          type: integer
        settings:
          $ref: '#/components/schemas/AgentSettings'
        stats:
          type: object
          properties:
            total_sessions:
              type: integer
            total_turns:
              type: integer
            avg_turns_per_session:
              type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AgentCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
        settings:
          $ref: '#/components/schemas/AgentSettings'
    AgentUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
        enabled:
          type: boolean
        settings:
          $ref: '#/components/schemas/AgentSettings'

    # Rules
...
    ErrorDetail:
      type: object
      required: [message]
      properties:
        field:
          type: string
        message:
          type: string
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
paths:
...
  /agents:
    get:
      summary: List agents
      tags: [Agents]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: enabled
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of agents
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Agent'
    post:
      summary: Create agent
      tags: [Agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreate'
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /agents/{agent_id}:
    parameters:
      - name: agent_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get agent
      tags: [Agents]
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update agent
      tags: [Agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentUpdate'
      responses:
        '200':
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
    delete:
      summary: Delete agent
      tags: [Agents]
      responses:
        '204':
          description: Agent deleted

  # Rules
...